requirements.txt
bs4
pypdf
ollama
joblib
chromadb
sqlparse
networkx
openpyxl
docx2txt
streamlit
python-pptx
unstructured
google-genai
scikit-learn
python-dotenv
langchain-chroma
langchain-ollama
langchain-community
google-generativeai
langchain-google-genai

-e .